Service mesh draft implementation to support clientRatelimit() and serviceRatelimit() filters with cluster information.
One disadvantage, probably not the only one, is that different cluster sizes will have buggy behaviour caused by the filter patching in the dataclient.
It would be nicer to have a way of getting the current ratelimit.Settings from the dataclient.
Maybe it is easier to get only the list of endpoints and communicate with some mesh protocol to distribute mesh relevant information.